refactor: restructure linting functionality to use internal library and improve error handling
