+++ internal/cmd/lint_runner.go
@@ -3,11 +3,11 @@ package cmd
 import (
 	"io"
 
+	"github.com/yoheimuta/protolint/internal/libinternal"
 	"github.com/yoheimuta/protolint/internal/osutil"
-	"github.com/yoheimuta/protolint/lib"
 )
 
-// CmdLintRunner implements the lib.LintRunner interface for cmd package
+// CmdLintRunner implements the LintRunner interface for cmd package
 type CmdLintRunner struct{}
 
 // NewCmdLintRunner creates a new CmdLintRunner
@@ -20,7 +20,7 @@ func (r *CmdLintRunner) Run(args []string, stdout, stderr io.Writer) osutil.Exit
 	return Do(args, stdout, stderr)
 }
 
-// Initialize registers the cmd lint runner with the lib package
+// Initialize registers the cmd lint runner with the internal library
 func Initialize() {
-	lib.SetLintRunner(NewCmdLintRunner())
+	libinternal.SetLintRunner(NewCmdLintRunner())
 }
